graphview bugfix: lineviews right constraint was wrong
