--- Minichallenge3/Views/GraphView/connections/GraphConnectionView.swift
@@ -1,10 +1,0 @@
-//
-//  GraphConnectionView.swift
-//  Minichallenge3
-//
-//  Created by Elias Paulino on 17/05/19.
-//  Copyright © 2019 Adriel Freire. All rights reserved.
-//
-
-import UIKit
-class GraphConnectionView: UIView { }
